resultats/repository/equipe-1-petits-farcis/back/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
geopy>=2.4.0
polyline>=2.0.0
pandas>=2.0.0
pillow>=10.0.0
